+++ test/all.test.js
@@ -15,5 +15,11 @@ chai.config.includeStack = true;
 
 // tests
 
+/* jshint expr: true */
+/* global describe, it */
+
 describe('Tests', function() {
+	it.skip('all features', function() {
+		expect(fs).to.be.ok;
+	});
 });